[hotfix][iteration] Add the flink-ml-tests project
